GPV-732 Added logic to improve 'skip' functionality
This tool relies on the presence or absence of certain
sentinel files to determine whether or not to execute a
step.  However, if a sentinel file is absent for a step,
yet the user wishes to skip that step, the old implementation
does not know to skip because of that absence.  Touching
the sentinel file will allow the skip to occur as expected.

Co-authored-by: Richard Gostanian <rgostanian@vmware.com>
Co-authored-by: Stu Pollock <stuartp@vmware.com>