GPV-936 Add logic to skip query generation for 07_multi_user
In the past, we always regenerate the queries for the 07_multi_user step
regardless we changed the CU or Scale. When we have a large scale database,
just re-generating the queries will take over an hour (for 1TB TPC-DS data).

We add a switch to the tpcds_variables.sh called RUN_QGEN.
If the RUN_QGEN is `true`, then we will regenerate the queries. Otherwise, we
just skip it.

There are two locations to generate the queries.
One is at the rollout.sh, where we use the `dsqgen` to generate the query
templates. The other one is at the test.sh, where we convert the query
templates into final queries.

The RUN_QGEN variable controls the skip logic in both locations.

Co-authored-by: Harshil Shah <sharshil@vmware.com>
Co-authored-by: Richard Gostanian <rgostanian@vmware.com>
Co-authored-by: Xin Zhang <zhxin@vmware.com>